Update link to CLJS JAR
ASCIIDoc does not have a mechanism for adding the `rel` attribute to
links. As a workaround, I used the passthrough syntax to add the
proper link relation.

The link itself points to the latest release page at GitHub. You can
make the link dynamic by adding a line of JavaScript to `footer.ftl`
from the current theme assets:

```
<script>
$(document).ready(function() {
    $.getJSON("https://api.github.com/repos/clojure/clojurescript/releases/latest").done(function(release) {var asset = release.assets[0].browser_download_url; $("a[rel='latest-version']").attr("href", asset)}) });
</script>
```