Make the set up clearer
I was baffled at first when I read the set up steps. It's also easier to point the user directly to the JAR.

Also I'd recommend to stick to one approach at a time: either JAR or CLJ - I recommend JAR, to keep things extremely simple for the user. Please let me know if you'd like help with that. It would make the tutorial much simpler.